requirements.txt
# Core dependencies
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Data processing
pdfplumber==0.11.9
python-docx>=1.0.0

# Vector database
chromadb>=0.4.0

# LLM clients
openai>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
python-multipart>=0.0.6
